CashUp_v2/apps/core/requirements.txt
# Core Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
aio-pika==9.4.0
aiohttp==3.9.1
feedparser==6.0.10
vaderSentiment==3.3.2
snownlp==0.12.3

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Logging and Monitoring
loguru==0.7.2
structlog==23.2.0

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0
pyyaml==6.0.1
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1